[SHARP] Added output of the current subscription
MOved ACR out of the way for now

These are all good commits to go to black

Commenting out docker install on jumphost

Install is SO SLOW, dont do it the crurriculum way for now

Installing project1 container

BUilding PUCE

Added another subscription

Added westeu webapp

This is all good, ready to build useast again

Only, this time, it will be forced into the IOT4 Subscription.

So this is where it will get built.

THis finishes the azure web app for week 13

Add all the github code in

Added all the TFC Code in