[ELAS] Packer AMI Built with 2 Disks
us-east-1 is running with 2 servers and with the packer golden ami
us-west-2 is up with 0 resources
eu-central-1 is up with 0 resources
azure week-12 stack is up and running on a single disk solution
azure week-13 is up for the students for now
elastio is deployed to the three regions

FInished the packer scrupt and move it up a notch

This gets a 12GB disk with index.html up in azure!

Ready to backup to elastio!

Finish